requirements.txt
faiss-cpu
google-genai==1.46.0
pypdf
numpy